# fly: improve performance of fly watch