## Changes proposed by this PR

Optimize TimestampedWriter for performance and reduced allocations. Added some tests as well.

Significantly improves performance of TimestampedWriter by implementing several optimizations:

- Uses a buffer pool to eliminate allocations between Write calls
- Processes input by segments between newlines rather than byte-by-byte
- Optimizes timestamp formatting with direct byte manipulation
- Pre-allocates and reuses buffers for newline tracking

Benchmark improvements:
- Eliminated allocations in most operations (0 allocs/op)
- Medium inputs with timestamps improved from 445.8 ns/op to 168.2 ns/op
- Large inputs with timestamps improved from 2284 ns/op to 871.2 ns/op
- 6x faster for SetTimestamp operations
- Maintains exact same output behavior

## Release Note

* Improve performance of `fly watch`